[02.06-안승우] 판매작품/작업해요 폴더명 변경
