+++ app/src/main/java/io/sinzak/android/ui/main/profile/art/adapter/Binding.kt
@@ -2,7 +2,7 @@ package io.sinzak.android.ui.main.profile.adapter
 
 import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.RecyclerView
-import io.sinzak.android.ui.main.profile.sale_with_work.adapter.SaleWorkAdapter
+import io.sinzak.android.ui.main.profile.art.adapter.SaleWorkAdapter
 
 @BindingAdapter("adapter")
 fun setAdapter(view : RecyclerView, adapter: SaleWorkAdapter)
