+++ app/src/main/res/layout/fragment_work.xml
@@ -5,15 +5,15 @@
     <data>
         <variable
             name="vm"
-            type="io.sinzak.android.ui.main.profile.sale_with_work.WorkViewModel" />
+            type="io.sinzak.android.ui.main.profile.art.WorkViewModel" />
         <variable
             name="fg"
-            type="io.sinzak.android.ui.main.profile.sale_with_work.WorkFragment" />
+            type="io.sinzak.android.ui.main.profile.art.WorkFragment" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.main.profile.sale_with_work.WorkFragment">
+        tools:context=".ui.main.profile.art.WorkFragment">
 
         <ImageView
             android:layout_width="24dp"
